requirements.txt
flask
redis
smooch
promise
PyJWT